requirements.txt
# Core Flask dependencies
Flask==3.0.3
requests==2.32.3
validators==0.22.0
python-dotenv==1.0.1

# ML/Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# AI Image Detection (Real vs AI-Generated)
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
datasets>=2.14.0
Pillow>=10.0.0
